Fix bookmarking in report

+++ src/dpr/routes/journeys/my-reports/bookmarks/controller.ts
@@ -14,15 +14,12 @@ class BookmarkController {
 
   POST: RequestHandler = async (req, res) => {
     const { dprUser } = LocalsHelper.getValues(res)
-    const { reportId, id, reportType } = req.body
-    await this.services.bookmarkService.addBookmark(dprUser.id, reportId, id, reportType)
-    res.end()
-  }
-
-  DELETE: RequestHandler = async (req, res) => {
-    const { dprUser } = LocalsHelper.getValues(res)
-    const { id, reportId } = req.body
-    await this.services.bookmarkService.removeBookmark(dprUser.id, id, reportId)
+    const { reportId, id, reportType, type } = req.body
+    if (type === 'add') {
+      await this.services.bookmarkService.addBookmark(dprUser.id, reportId, id, reportType)
+    } else {
+      await this.services.bookmarkService.removeBookmark(dprUser.id, id, reportId)
+    }
     res.end()
   }
 }
